Logic to send test slack notification and check response

+++ GlasgowAstro.GuideAlert/Models/GuideAlertSettings.cs
@@ -10,6 +10,6 @@
 
         public string PhdHost { get; set; }
 
-        public int PhdPort { get; set; }
+        public ushort PhdPort { get; set; }
     }
 }
